feat!: Remove SNS notification on error
Added in #12, a notification would be sent to SNS when the doc tool failed.

The SNS topic has one subscriber - email, to a colleague who left a few years ago.

In this change, we remove this error notification as we're never receiving the emails.
The errors are still shown in the log page.